% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArtificialExtinction.R
\name{ArtificialExtinction}
\alias{ArtificialExtinction}
\alias{ArtificialExtinction.matrix}
\alias{ArtificialExtinction.phyDat}
\alias{ArtEx}
\title{Artificial Extinction}
\usage{
ArtificialExtinction(
  dataset,
  template,
  subject,
  replacement = "?",
  sampleFrom = NULL
)

\method{ArtificialExtinction}{matrix}(
  dataset,
  template,
  subject,
  replacement = "?",
  sampleFrom = NULL
)

\method{ArtificialExtinction}{phyDat}(
  dataset,
  template,
  subject,
  replacement = "?",
  sampleFrom = NULL
)

ArtEx(dataset, template, subject, replacement = "?", sampleFrom = NULL)
}
\arguments{
\item{dataset}{Phylogenetic dataset of class \code{phyDat} or \code{matrix}.}

\item{template}{Taxon to use as a template.}

\item{subject}{Vector identifying subject taxa, by name or index.}

\item{replacement}{Character specifying whether deleted tokens should be
replaced with:
\itemize{
\item \verb{?}: The ambiguous token, \verb{?};
\item \code{binary}: The tokens \code{0} or \code{1}, with equal probability;
\item \code{uniform}: One of the tokens present in \code{sampleFrom}, with equal
probability;
\item \code{sample}: One of the tokens present in \code{sampleFrom}, sampled according
to their frequency.
}}
}
\value{
A dataset with the same class as \code{dataset} in which entries that
are ambiguous in \code{template} are made ambiguous in \code{subject}.
}
\description{
Remove tokens that do not occur in a fossil 'template' taxon from a living
taxon, to simulate the process of fossilization in removing data from
a phylogenetic dataset.
}
\details{
Note: this simple implementation does not account for character contingency,
e.g. characters whose absence imposes inapplicable or absent tokens on
dependent characters.
}
\examples{
set.seed(1)
dataset <- matrix(c(sample(0:2, 4 * 8, TRUE),
                    '0', '0', rep('?', 6)), nrow = 5,
                    dimnames = list(c(LETTERS[1:4], 'FOSSIL'),
                                    paste('char', 1:8)), byrow = TRUE)
artex <- ArtificialExtinction(dataset, 'FOSSIL', c('A', 'C'))
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
