% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_write.R
\name{as.Newick}
\alias{as.Newick}
\alias{as.Newick.phylo}
\alias{as.Newick.list}
\alias{as.Newick.multiPhylo}
\title{Write a phylogenetic tree in Newick format}
\usage{
as.Newick(x)

\method{as.Newick}{phylo}(x)

\method{as.Newick}{list}(x)

\method{as.Newick}{multiPhylo}(x)
}
\arguments{
\item{x}{Object to convert to Newick format.  See below for supported classes.}
}
\value{
\code{as.Newick} returns a character string representing \code{tree} in Newick
format.
}
\description{
Creates a character string describing a phylogenetic tree in Newick format,
using R's internal tip numbering.
}
\seealso{
\itemize{
\item \code{\link{RenumberTips}}
\item \code{\link[ape:write.tree]{ape::write.tree}}
}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
