% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Splits.R
\name{as.Splits}
\alias{as.Splits}
\title{As splits}
\usage{
as.Splits(x, tipLabels = NULL, ...)
}
\arguments{
\item{tipLabels}{Character vector specifying sequence in which to order
tip labels.  Label order must (currently) match to combine or compare separate
\code{Splits} objects.}

\item{tr}{A tree of class \code{\link[ape:read.tree]{phylo}}.}

\item{asSplits}{Logical specifying whether to return a \code{Splits} object,
or an unannotated two-dimensional array (useful where performance is
paramount).}
}
\value{
Returns an object of class \code{Splits}, or (if \code{asSplits = FALSE}) a
two-dimensional array of 32-bit integers, which each bit specifying whether
a tip is a member of the split.  Splits are named according to the node
that defines them.
}
\description{
Converts a phylogenetic tree to an array of bipartition splits.
}
\examples{
as.Splits(ape::rtree(6, tip.label=1:6, br=NULL))

}
\author{
Martin R. Smith
}
