% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brewer}
\alias{brewer}
\title{Brewer palettes}
\format{
An object of class \code{list} of length 12.
}
\source{
{
\itemize{
\item \href{http://colorbrewer2.org/#type=diverging&scheme=RdYlBu&n=3}{ColourBrewer2.org}
\item \href{http://mkweb.bcgsc.ca/colorblind/}{Martin Krzywinski}
}
}
}
\usage{
brewer
}
\description{
A list of eleven Brewer palettes containing one to eleven colours that
are readily distinguished by colourblind viewers, followed by a twelfth
12-colour palette adapted for colour blindness.
}
\examples{
data("brewer", package="TreeTools")
plot(0, type='n', xlim=c(1, 12), ylim=c(12, 1),
     xlab = 'Colour', ylab='Palette')
for (i in seq_along(brewer)) text(seq_len(i), i, col=brewer[[i]])

}
\keyword{datasets}
