% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_information.R
\name{CladisticInfo}
\alias{CladisticInfo}
\alias{PhylogeneticInfo}
\alias{PhylogeneticInformation}
\alias{CladisticInformation}
\title{Cladistic information content}
\usage{
CladisticInfo(x)

PhylogeneticInfo(x)

PhylogeneticInformation(x)

CladisticInformation(x)
}
\arguments{
\item{x}{Tree of class \code{phylo}, or a list thereof.}
}
\value{
Returns the phylogenetic or clustering information content
of the input tree(s).
}
\description{
Calculate the cladistic (phylogenetic) information content of a
phylogenetic object, \emph{sensu} Thorley \emph{et al.} (1998).
}
\details{
The CIC is the logarithm of the number of binary trees that include the
specified topology.  A base two logarithm gives an information content in
bits.

The CIC was originally proposed by Rohlf (1982), and formalised as CIC, with
an information-theoretic justification, by Thorley \emph{et al}. (1998).
Steel and Penny (2006) term the equivalent quantity 'phylogenetic information
content' in the context of individual characters.

The number of binary trees consistent with a cladogram provides a more
satisfactory measure of the resolution of a tree than simply
counting the number of edges resolved (Page, 1992).
}
\references{
\insertRef{Page1992}{TreeTools}

\insertRef{Rohlf1982}{TreeTools}

\insertRef{Steel2006}{TreeTools}

\insertRef{Thorley1998}{TreeTools}
}
\seealso{
Other tree information functions: 
\code{\link{TreesMatchingTree}()}

Other tree characterization functions: 
\code{\link{TotalCopheneticIndex}()}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree characterization functions}
\concept{tree information functions}
