% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_information.R
\name{ClusterInfo}
\alias{ClusterInfo}
\title{Clustering Information}
\usage{
ClusterInfo(x)
}
\arguments{
\item{x}{Tree of class \code{phylo}, or a list thereof.}
}
\description{
#TODO
}
\details{
update TreeDist docs to refer to this function.

Unbalanced trees contain more uneven clusters, and thus contain
more clustering information than balanced trees.

This is considered a defect by Rohlf (1982) and Page (1992), on the basis
that tree shape does not influence the amount of \emph{cladistic} information
held within a tree.  Cladistic information sees the purpose of trees as
depicting the evolutionary relationships of a set of taxa.  On this view,
a tree is either 'right' or 'wrong'.

Clustering information represents a different concept of the information
within a tree; trees serve to define groups of closely related taxa.
On this view, an unalanced tree makes more grouping ('nesting') statements
than a balanced one (Adams, 1986), corresponding to more information.
As such, a tree can differ from the 'true' tree in small details, yet
still convey much 'true' information.
}
\examples{
tree <- CollapseNode(BalancedTree(10), c(12:13, 19))
plot(tree)
nodelabels()
ClusterInfo(tree)
ClusterInfo(BalancedTree(8))
ClusterInfo(PectinateTree(8))


}
\references{
\insertRef{Adams1986}{TreeTools}

\insertRef{Page1992}{TreeTools}

\insertRef{Rohlf1982}{TreeTools}
}
\seealso{
Other tree information functions: 
\code{\link{PhylogeneticInfo}()},
\code{\link{TreesMatchingTree}()}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree information functions}
