% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_information.R
\name{ClusteringInfo}
\alias{ClusteringInfo}
\title{Clustering Information}
\usage{
ClusteringInfo(x, nTip = NTip(x))
}
\arguments{
\item{x}{Tree of class \code{phylo}, or a list thereof.}
}
\description{
Clustering Information
}
\seealso{
Other tree information functions: 
\code{\link{PhylogeneticInfo}()},
\code{\link{TreesMatchingTree}()}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree information functions}
