% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitFunctions.R
\name{CompatibleSplits}
\alias{CompatibleSplits}
\alias{.CompatibleSplit}
\title{Which splits are compatible?}
\usage{
CompatibleSplits(splits, splits2)

.CompatibleSplit(a, b, nTip)
}
\arguments{
\item{splits}{An object of class \link{as.Splits}{`Splits`}.}

\item{splits2}{A second \code{Splits} object.}

\item{a, b}{Integer representations of splits, from an row of a \code{Splits} object}
}
\value{
A logical matrix specifying whether each split in \code{splits} is
compatible with each split in \code{splits2}.

logical vector stating whether splits are compatible
}
\description{
Which splits are compatible?
}
\section{Functions}{
\itemize{
\item \code{.CompatibleSplit}: Evaluate a single split pair
}}

\examples{

splits <- as.Splits(BalancedTree(8))
splits2 <- as.Splits(PectinateTree(8))

summary(splits)
summary(splits2)

CompatibleSplits(splits, splits2)

}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\keyword{internal}
