% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Splits.R
\name{.DecodeRaw}
\alias{.DecodeRaw}
\alias{.DecodeLastRaw}
\alias{Decoders}
\alias{.DecodeBinary}
\title{Decode Splits objects}
\usage{
.DecodeRaw(n, stopAt = 8L, print = FALSE, appendLF = FALSE)

.DecodeLastRaw(n, nTip, ...)

.DecodeBinary(n, nTip, print = FALSE, appendLF = FALSE, ...)
}
\arguments{
\item{n}{Number to decode.}

\item{stopAt}{Integer specifying number of tips in partial raw element.}

\item{print}{Logical specifying whether output is to be printed.}

\item{appendLF}{Logical specifying whether to append line feed character to
output.}

\item{nTip}{Integer specifying number of tips in original tree.}
}
\value{
Return a human-readable representation of split content.
}
\description{
Internal functions to decode raw representation of splits.
}
\keyword{internal}
