% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decompose.R
\name{Decompose}
\alias{Decompose}
\title{Decompose additive (ordered) phylogenetic characters}
\usage{
Decompose(dataset, indices)
}
\arguments{
\item{dataset}{A phylogenetic data matrix of \pkg{phangorn} class \code{phyDat},
whose names correspond to the labels of any accompanying tree.}

\item{indices}{Integer or logical vector specifying indices of characters
that should be decomposed}
}
\value{
\code{Decompose()} returns a \code{phyDat} object in which the specified
characters have been decomposed into a number of binary characters.
}
\description{
\code{Decompose()} decomposes additive characters into a series of binary
characters, which is mathematically equivalent when analysed under
equal weights parsimony.  (This equivalence is not exact
under implied weights or under probabilistic tree inference methods.)
}
\examples{
data("Lobo")

# Identify character 11 as additive
# Character 11 will be replaced with two characters
# The present codings 0, 1 and 2 will be replaced with 00, 10, and 11.
decomposed <- Decompose(Lobo.phy, 11)

attr(Lobo.phy, "nr")   # 113 characters
attr(decomposed, "nr") # 114 character rows

}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
