% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Consensus.R
\name{.CountSplits}
\alias{.CountSplits}
\title{Count splits and calculate information content}
\usage{
.CountSplits(trees)
}
\arguments{
\item{trees}{List or multiPhylo object containing phylogenetic trees of
class "phylo", on the same set of tips.}
}
\value{
\code{.CountSplits()} currently returns a list with four elements:
\itemize{
\item \code{splits} is a logical matrix; each column corresponds to a leaf, each row
to a split
\item \code{count} lists the number of occurrences of that split
\item \code{pic} lists the phylogenetic information content of that split, assuming
that split frequency is proportional to split probability
\item \code{cic} lists the clustering information content of each split, on the same
basis.
Splits are listed tree by tree, including duplicates.
The order of splits corresponds to their position in the cluster table
representation of each tree, so may differ from the sequence given by
\code{as.Splits(trees)}.
}
}
\description{
Counts the number of times that each split occurs in a posterior sample of
trees, and computes the information content implied by split membership
and probability.
}
\details{
This is marked as an internal function as its behaviour is subject to change;
it may be replaced with more efficient functions for specific applications
in future. Please alert the maintainer if you plan to call the function
directly in your own work.
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\keyword{internal}
