% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_files.R
\name{EndSentence}
\alias{EndSentence}
\title{Add full stop to end of a sentence}
\usage{
EndSentence(string)
}
\arguments{
\item{string}{Input string}
}
\value{
\code{EndSentence()} returns \code{string}, punctuated with a final full stop
(period).`
}
\description{
Add full stop to end of a sentence
}
\examples{
EndSentence("Hello World") # "Hello World."
}
\seealso{
Other string parsing functions: 
\code{\link{MatchStrings}()},
\code{\link{MorphoBankDecode}()},
\code{\link{RightmostCharacter}()},
\code{\link{Unquote}()}
}
\author{
Martin R. Smith
}
\concept{string parsing functions}
