% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_information.R
\name{FrequencyToEntropy}
\alias{FrequencyToEntropy}
\title{Entropy of a table of frequencies}
\usage{
FrequencyToEntropy(...)
}
\arguments{
\item{\dots}{Frequencies  of observations}
}
\value{
\code{FrequencyToEntropy()} returns the entropy of the input observations.
}
\description{
Entropy of a table of frequencies
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
