% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_generation.R
\name{GenerateTree}
\alias{GenerateTree}
\alias{RandomTree}
\alias{PectinateTree}
\alias{BalancedTree}
\alias{StarTree}
\title{Generate pectinate, balanced or random trees}
\usage{
RandomTree(tips, root = FALSE)

PectinateTree(tips)

BalancedTree(tips)

StarTree(tips)
}
\arguments{
\item{tips}{An integer specifying the number of tips, or a character vector
naming the tips, or any other object from which \code{\link[=TipLabels]{TipLabels()}} can
extract leaf labels.}

\item{root}{Character or integer specifying tip to use as root, if desired;
or \code{FALSE} for an unrooted tree.}
}
\value{
Each function returns an unweighted binary tree of class \code{phylo} with
the specified leaf labels. Trees are rooted unless \code{root = FALSE}.

\code{RandomTree()} returns a random topology, optionally rooting the
tree on a given tip.

\code{PectinateTree()} returns a pectinate (caterpillar) tree.

\code{BalancedTree()} returns a balanced (symmetrical) tree.

\code{StarTree()} returns a completely unresolved (star) tree.
}
\description{
\code{RandomTree()}, \code{PectinateTree()}, \code{BalancedTree()} and \code{StarTree()}
generate trees with the specified shapes and leaf labels.
}
\examples{
RandomTree(LETTERS[1:10])

data('Lobo')
RandomTree(Lobo.phy)

plot(PectinateTree(LETTERS[1:10]))

plot(BalancedTree(LETTERS[1:10]))
plot(StarTree(LETTERS[1:10]))

}
\seealso{
Other tree generation functions: 
\code{\link{NJTree}()},
\code{\link{SingleTaxonTree}()},
\code{\link{TreeNumber}}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree generation functions}
