% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_generation.R
\name{GetTipNames}
\alias{GetTipNames}
\title{Extract tip names from a dataset of unknown class}
\usage{
GetTipNames(tips)
}
\arguments{
\item{tips}{An integer specifying the number of tips, or a character vector
naming the tips, or a tree of class phylo, or a phylogenetic dataset whose
names will be used to name the tips.}
}
\value{
Character vector listing tip names.
}
\description{
Extract tip names from a dataset of unknown class
}
\author{
Martin R. Smith
}
\keyword{internal}
