% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_generation.R
\name{Hamming}
\alias{Hamming}
\title{Hamming distance between taxa in a phylogenetic dataset}
\usage{
Hamming(dataset, ratio = TRUE)
}
\arguments{
\item{dataset}{object of class phyDat}

\item{ratio}{Logical specifying whether to weight distance against
maximum possible.}
}
\value{
\code{Hamming()} returns an object of class \code{dist} listing the Hamming
distance between each pair of taxa.
}
\description{
The Hamming distance between a pair of taxa is the number of characters
with a different coding, i.e. the smallest number of evolutionary steps
that must have occurred since their common ancestor.
}
\details{
Tokens that contain the inapplicable state are treated as requiring no steps
to transform into any applicable token.
}
\seealso{
Used to construct neighbour joining trees in \code{\link[=NJTree]{NJTree()}}.

\code{dist.hamming()} in the 'phangorn' package provides an alternative
implementation.
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
