% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeNumber.R
\name{is.TreeNumber}
\alias{is.TreeNumber}
\title{Is an object a \code{TreeNumber} object?}
\usage{
is.TreeNumber(x)
}
\arguments{
\item{x}{R object.}
}
\value{
\code{is.TreeNumber()} returns a logical vector of length one specifying
whether \code{x} inherits the class \code{"TreeNumber"}.
}
\description{
Is an object a \code{TreeNumber} object?
}
\examples{
is.TreeNumber(FALSE) # FALSE 
is.TreeNumber(as.TreeNumber(BalancedTree(5))) # TRUE
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
