% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUtreebalance.R
\name{J1Index}
\alias{J1Index}
\title{Robust universal tree balance index}
\usage{
J1Index(tree, q = 1, nonrootdomfactor = FALSE)
}
\description{
Calculate tree balance index J^1 (when nonrootdomfactor = FALSE) or
J^{1c} (when nonrootdomfactor = TRUE) from \insertRef{Lemant2022}{TreeTools}.
If population sizes are missing then the function assigns
size 0 to internal nodes, and size 1 to leaves.
}
\examples{
tree1 <- data.frame(Parent = c(1,1,1,1,2,3,4),
                    Identity = 1:7,
                    Population = c(1, rep(5, 6)))
J1Index(tree1)
tree2 <- data.frame(Parent = c(1,1,1,1,2,3,4),
                    Identity = 1:7,
                    Population = c(rep(0, 4), rep(1, 3)))
J1Index(tree2)
tree3 <- data.frame(Parent = c(1,1,1,1,2,3,4),
                    Identity = 1:7,
                    Population = c(0, rep(1, 3), rep(0, 3)))
J1Index(tree3)
cat_tree <- data.frame(Parent = c(1, 1:14, 1:15, 15),
                       Identity = 1:31,
                       Population = c(rep(0, 15), rep(1, 16)))
J1Index(cat_tree)
sym_tree <- data.frame(Parent = c(1, rep(1:15, each = 2)),
                       Identity = 1:31,
                       Population = c(rep(0, 15), rep(1, 16)))
J1Index(sym_tree)
}
\references{
\insertAllCited{}
}
\seealso{
Other tree characterization functions: 
\code{\link{CladisticInfo}()},
\code{\link{Consensus}()},
\code{\link{Stemwardness}},
\code{\link{TotalCopheneticIndex}()}
}
\author{
Rob Noble, adapted by Martin R. Smith
}
\concept{tree characterization functions}
