% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports-manual.R
\name{keep_tip}
\alias{keep_tip}
\title{C++ wrapper for keep_tip}
\usage{
keep_tip(edge, keep)
}
\arguments{
\item{edge}{Tree of class \code{phylo}, in (loose) preorder.}

\item{keep}{Logical vector specifying whether to keep each leaf.}
}
\description{
Wrapper for C++ function \code{keep_tip()}.
No attempt is made to validate input; specifying invalid input may crash
R.
Behaviour is subject to change without notice.
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
