% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.R
\name{ListAncestors}
\alias{ListAncestors}
\title{List ancestors}
\usage{
ListAncestors(parent, child, node)
}
\arguments{
\item{parent}{the 'parent' column of the edges property of a tree of class \code{phylo};}

\item{child}{the 'child' column of the edges property of a tree of class \code{phylo};}

\item{node}{the number of the node or tip whose ancestors are required.}
}
\value{
\code{ListAncestors()} returns a vector of the numbers of the nodes
ancestral to the given \code{node}, including the root node.
}
\description{
Reports all ancestors of a given node
}
\details{
To observe the number of a node or tip, use
\code{plot(tree); \link[ape]{nodelabels}(); \link[ape:nodelabels]{tiplabels}();}
}
\examples{
tree   <- ape::read.tree(text='(1, (2, (3, (4, 5))));')
edge   <- tree$edge
parent <- tree$edge[, 1]
child  <- tree$edge[, 2]
ListAncestors(parent, child, 4L)

}
\seealso{
\itemize{
\item \code{phangorn:::Ancestors}, a less efficient implementation on which this
code is based.
}

Other tree navigation: 
\code{\link{AllAncestors}()},
\code{\link{AncestorEdge}()},
\code{\link{CladeSizes}()},
\code{\link{DescendantEdges}()},
\code{\link{EdgeAncestry}()},
\code{\link{EdgeDistances}()},
\code{\link{MRCA}()},
\code{\link{NDescendants}()},
\code{\link{NodeDepth}()},
\code{\link{NodeOrder}()},
\code{\link{NonDuplicateRoot}()},
\code{\link{RootNode}()}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree navigation}
