% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_rearrangement.R
\name{MakeTreeBinary}
\alias{MakeTreeBinary}
\title{Generate binary tree by collapsing polytomies}
\usage{
MakeTreeBinary(tree)
}
\arguments{
\item{tree}{A tree of class \code{\link[ape:read.tree]{phylo}}.}
}
\value{
\code{MakeTreeBinary()} returns a rooted binary tree of class \code{phylo},
corresponding to tree uniformly selected from all those compatible with
the input tree topologies.
}
\description{
\code{MakeTreeBinary()} resolves, at random, all polytomies in a tree or set of
trees, such that all trees compatible with the input topology are drawn
with equal probability.
}
\examples{
MakeTreeBinary(CollapseNode(PectinateTree(7), c(9, 11, 13)))
Unroot(MakeTreeBinary(StarTree(5)))
}
\seealso{
Resolve polytomies such that some resolutions are more probable
than others using \code{\link[ape:multi2di]{ape::multi2di()}}.
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
