% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchNodes.R
\name{MatchEdges}
\alias{MatchEdges}
\alias{MatchNodes}
\title{Match nodes and edges between trees}
\usage{
MatchEdges(x, table, nomatch = NA_integer_)

MatchNodes(x, table, nomatch = NA_integer_, tips = FALSE)
}
\arguments{
\item{x}{Tree whose nodes are to be matched.}

\item{table}{Tree containing nodes to be matched against.}

\item{nomatch}{Integer value that will be used in place of \code{NA} in the case
where no match is found.}

\item{tips}{Logical specifying whether to return matches for tips;
unless \code{TRUE}, only the matches for internal nodes will be returned.}
}
\description{
\code{MatchNodes()} and \code{MatchEdges()} matches nodes or edges in one tree to
entries in the second that denote a clade with identical tip labels.
}
\details{
The current implementation is potentially inefficient.
Please contact the maintainer to request a more efficient implementation if
this function is proving a bottleneck.
}
\examples{
MatchNodes(BalancedTree(8), RootTree(BalancedTree(8)))
}
\seealso{
Other tree navigation:
\code{\link[=AncestorEdge]{AncestorEdge()}},
\code{\link[=CladeSizes]{CladeSizes()}},
\code{\link[=DescendantEdges]{DescendantEdges()}},
\code{\link[=EdgeAncestry]{EdgeAncestry()}},
\code{\link[=EdgeDistances]{EdgeDistances()}},
\code{\link[=ListAncestors]{ListAncestors()}},
\code{\link[=MRCA]{MRCA()}},
\code{\link[=NDescendants]{NDescendants()}},
\code{\link[=NodeDepth]{NodeDepth()}},
\code{\link[=NodeNumbers]{NodeNumbers()}},
\code{\link[=NodeOrder]{NodeOrder()}},
\code{\link[=PaintTree]{PaintTree()}},
\code{\link[=RootNode]{RootNode()}}

Other tree properties:
\code{\link[=Cherries]{Cherries()}},
\code{\link[=ConsensusWithout]{ConsensusWithout()}},
\code{\link[=EdgeRatio]{EdgeRatio()}},
\code{\link[=LongBranch]{LongBranch()}},
\code{\link[=NSplits]{NSplits()}},
\code{\link[=NTip]{NTip()}},
\code{\link[=NodeNumbers]{NodeNumbers()}},
\code{\link[=PathLengths]{PathLengths()}},
\code{\link[=SplitsInBinaryTree]{SplitsInBinaryTree()}},
\code{\link[=TipLabels]{TipLabels()}},
\code{\link[=TreeIsRooted]{TreeIsRooted()}},
\code{\link[=Treeness]{Treeness()}}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree navigation}
\concept{tree properties}
