% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckTips.R
\name{MatchStrings}
\alias{MatchStrings}
\title{Check for mismatch between tip labels}
\usage{
MatchStrings(x, table, Fail = stop, max.distance = 0.5, ...)
}
\arguments{
\item{x, table}{Character vectors, in which all elements of \code{x} are expected
to occur in \code{table}.}

\item{Fail}{Function to call if a mismatch is found.}

\item{\dots}{Additional arguments to \code{\link[base:agrep]{agrep()}}.}
}
\description{
Checks that entries in one character vector occur in another, suggesting
corrections for mismatched elements.
}
