% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_files.R
\name{MorphoBankDecode}
\alias{MorphoBankDecode}
\title{Decode MorphoBank text}
\usage{
MorphoBankDecode(string)
}
\arguments{
\item{string}{String to process}
}
\value{
A string with new lines and punctuation reformatted
}
\description{
Converts strings from MorphoBank notes into a format compatible with Latex / bookdown
}
\author{
Martin R. Smith
}
