% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combinatorics.R
\name{N1Spr}
\alias{N1Spr}
\title{Number of trees one SPR step away
Formula given by Allen and Steel (2001).}
\usage{
N1Spr(n)
}
\arguments{
\item{n}{Number of tips in tree.}
}
\description{
Number of trees one SPR step away
Formula given by Allen and Steel (2001).
}
\references{
\insertRef{Allen2001}{TreeTrunk}
}
