% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_numbering.R
\name{NeworderPhylo}
\alias{NeworderPhylo}
\title{neworder_phylo
Wrapper for the ape function}
\usage{
NeworderPhylo(nTaxa, parent, child, nb.edge, whichwise)
}
\description{
neworder_phylo
Wrapper for the ape function
}
\seealso{
Other C wrappers: \code{\link{Cladewise}},
  \code{\link{NeworderPruningwise}},
  \code{\link{OrderEdgesNumberNodes}},
  \code{\link{RenumberTree}}
}
\concept{C wrappers}
\keyword{internal}
