% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_properties.R
\name{NodeDepth}
\alias{NodeDepth}
\title{Distance of each node from tree exterior}
\usage{
NodeDepth(x, includeTips = TRUE)
}
\arguments{
\item{x}{A tree of class \code{phylo}, its \verb{$edge} property, or a list thereof.}

\item{includeTips}{Logical specifying whether to include leaves
(each of depth zero) in return value.}
}
\value{
\code{NodeDepth()} returns an integer vector specifying the depth of
each external and internal node in \code{x}.
}
\description{
\code{NodeDepth()} evaluates how 'deep' each node is within an unrooted tree.
}
\details{
A leaf is considered to represent the surface of the tree and has depth zero.
The depth of an edge is the smallest of the depths of the two vertices that
it connects.
The depth of a vertex is one greater than its second deepest incident edge.
(The deepest incident edge leads 'deeper' into the tree.)
}
\examples{
tree <- CollapseNode(BalancedTree(10), c(12:13, 19))
plot(tree)
nodelabels(NodeDepth(tree, FALSE))


}
\seealso{
\code{\link[ape:node.depth]{ape::node.depth}} returns the number of tips descended from a
node.

Other tree navigation: 
\code{\link{AllAncestors}()},
\code{\link{AncestorEdge}()},
\code{\link{DescendantEdges}()},
\code{\link{EdgeAncestry}()},
\code{\link{EdgeDistances}()},
\code{\link{MRCA}()},
\code{\link{NDescendants}()},
\code{\link{NodeOrder}()},
\code{\link{NonDuplicateRoot}()},
\code{\link{RootNode}()}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree navigation}
