% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nRootedShapes}
\alias{nRootedShapes}
\title{Number of rooted shapes}
\format{
An object of class \code{integer64} of length 55.
}
\source{
Wedderburn–Etherington numbers,
\href{https://oeis.org/A001190/b001190.txt}{OEIS A001190}
}
\usage{
nRootedShapes
}
\description{
\code{nRootedShapes} gives the nUmber of binary rooted trees on \emph{n} unlabeled
leaves.
}
\keyword{datasets}
