% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_properties.R
\name{NSplits}
\alias{NSplits}
\alias{NPartitions}
\title{Number of distinct partitions in a tree}
\usage{
NSplits(x)

NPartitions(x)
}
\arguments{
\item{x}{A phylogenetic tree of class \code{phylo}, or a list of such trees
(of class \code{list} or \code{multiPhylo}), or a \code{Splits} object,
or a vector of integers.}
}
\value{
Integer specifying the number of partitions in the specified trees,
or in a rooted tree with \code{n} tips.
}
\description{
Number of distinct partitions in a tree
}
\examples{
{
NPartitions(8)
NPartitions(PectinateTree(8))
NPartitions(as.Splits(BalancedTree(8)))
}
}
\author{
Martin R. Smith
}
