% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_generation.R
\name{PectinateTree}
\alias{PectinateTree}
\title{Generate a Pectinate Tree}
\usage{
PectinateTree(tips)
}
\arguments{
\item{tips}{An integer specifying the number of tips, or a character vector
naming the tips, or a dataset phylogenetic dataset whose names will be used
to name the tips.}
}
\value{
A tree of class \code{phylo}.
}
\description{
Generates a pectinate (caterpillar) tree with the specified tips
}
\seealso{
Other tree generation functions: \code{\link{NJTree}},
  \code{\link{RandomTree}}, \code{\link{SingleTaxonTree}}
}
\author{
Martin R. Smith
}
\concept{tree generation functions}
