% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_information.R
\name{PhylogeneticInfo}
\alias{PhylogeneticInfo}
\alias{PhylogeneticInformation}
\title{Cladistic information content}
\usage{
PhylogeneticInfo(x)

PhylogeneticInformation(x)
}
\arguments{
\item{x}{Tree of class \code{phylo}, or a list thereof.}
}
\value{
Returns the phylogenetic or clustering information content
of the input tree(s).
}
\description{
Calculate the cladistic (phylogenetic) information content of a
phylogenetic object, \emph{sensu} Thorley \emph{et al.} (1998).
}
\details{
#TODO
\itemize{
\item Splits to follow
\item Further information to follow
\item See also Clustering Information
}
}
\references{
\insertRef{Thorley1998}{TreeTools}
}
\seealso{
Other tree information functions: 
\code{\link{ClusteringInfo}()},
\code{\link{TreesMatchingTree}()}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree information functions}
