% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_files.R
\name{PhyToString}
\alias{PhyToString}
\alias{PhyDatToString}
\alias{PhydatToString}
\title{Extract character data from a phyDat object as a string}
\usage{
PhyToString(phy, parentheses = "{", collapse = "", ps = "",
  useIndex = TRUE, byTaxon = TRUE, concatenate = TRUE)

PhyDatToString(phy, parentheses = "{", collapse = "", ps = "",
  useIndex = TRUE, byTaxon = TRUE, concatenate = TRUE)

PhydatToString(phy, parentheses = "{", collapse = "", ps = "",
  useIndex = TRUE, byTaxon = TRUE, concatenate = TRUE)
}
\arguments{
\item{phy}{An object of class \code{\link{phyDat}}}

\item{parentheses}{Character specifying format of parentheses with which
to surround ambiguous tokens.  Choose from: \code{\{} (default), \code{[}, \code{(}, \code{<}.}

\item{collapse}{Character specifying text, perhaps \code{,}, with which to
separate multiple tokens within parentheses}

\item{ps}{Character specifying text, perhaps \code{;}, to append to the end of the string}

\item{useIndex}{(default: TRUE) Print duplicate characters multiple
times, as they appeared in the original matrix}

\item{byTaxon}{If TRUE, write one taxon followed by the next.
If FALSE, write one character followed by the next.}

\item{concatenate}{Logical specifying whether to concatenate all characters/taxa
into a single string, or to return a separate string
for each entry.}
}
\description{
Extract character data from a phyDat object as a string
}
\author{
Martin R. Smith
}
