% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Splits.R
\name{PolarizeSplits}
\alias{PolarizeSplits}
\title{Polarize splits on a single taxon}
\usage{
PolarizeSplits(x, pole = 1L)
}
\arguments{
\item{x}{Object that can be coerced into class \code{\link{Splits}}.}

\item{pole}{Numeric, character or logical vector identifying tip that will
polarize each split.}
}
\value{
\code{PolarizeSplits()} returns a \code{Splits} object in which \code{pole} is
represented by a zero bit
}
\description{
Polarize splits on a single taxon
}
\seealso{
Other Splits operations:
\code{\link[=LabelSplits]{LabelSplits()}},
\code{\link[=NSplits]{NSplits()}},
\code{\link[=NTip]{NTip()}},
\code{\link[=SplitFrequency]{SplitFrequency()}},
\code{\link{Splits}},
\code{\link[=SplitsInBinaryTree]{SplitsInBinaryTree()}},
\code{\link[=TipLabels]{TipLabels()}},
\code{\link[=TipsInSplits]{TipsInSplits()}},
\code{\link{match,Splits,Splits-method}},
\code{\link[=xor]{xor()}}
}
\concept{Splits operations}
