% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_files.R
\name{ReadTntTree}
\alias{ReadTntTree}
\alias{TNTText2Tree}
\title{Parse TNT Tree}
\usage{
ReadTntTree(filename, relativePath = NULL, keepEnd = 1L,
  tipLabels = NULL)

TNTText2Tree(treeText)
}
\arguments{
\item{filename}{character string specifying path to TNT \code{.tre} file.}

\item{relativePath}{(optional) character string specifying location of the
matrix file used to generate the TNT results, relative to the current working
directory, for portability.  Taxon names will be read from this file if they
are not specified by \code{tipLabels}.}

\item{keepEnd}{(optional, default 1) integer specifying how many elements of the file
path to conserve when creating relative path (see examples).}

\item{tipLabels}{(optional) character vector specifying the names of the
taxa, in the sequence that they appear in the TNT file.  If not specified,
taxon names will be loaded from the data file linked in the first line of the
\code{.tre} file specified in \code{filename}.}

\item{treeText}{Character string describing a tree, in the parenthetical
format output by TNT.}
}
\value{
a tree of class \code{phylo}.
}
\description{
Reads a tree from TNT's parenthetical output.
}
\details{
Supports trees that have been written from TNT in parenthetical notation
using \code{tsav*}.  If taxa have been saved using their names (\code{taxname=}),
then the tip labels will be read directly from the TNT \code{.tre} file.  If
taxa have been saved using their numbers (\code{taxname-}), tip labels will be
imported from the linked matrix file that is listed in the first line of the
\code{.tre} file.  Ensure that this file exists in the expected location -- if
it does not, use the \code{relativePath} argument to override this default, or
specify \code{tipLabels} to set manually.
}
\section{Functions}{
\itemize{
\item \code{TNTText2Tree}: Converts text representation of a tree in TNT to an
object of class \code{phylo}.
}}

\examples{
  # In the examples below, TNT has read a matrix from
  # "c:/TreeTools/input/dataset.nex"
  # The results of an analysis were written to
  # "c:/TreeTools/output/results1.tnt"
  #
  # results1.tnt will contain a hard-coded reference to
  # "c:/TreeTools/input/dataset.nex".

  # On the original machine, it would be possible to read this hard-coded
  # reference from results.tnt:

  \dontrun{
  # Works on original machine but not elsewhere:
  ReadTntTree('output/results1.tnt')

  # Takes only the filename from the results
  ReadTntTree('output/results1.tnt', 'input/dataset.nex')
  }

  # These datasets are provided with the `TreeTools` package, which will
  # probably not be located at c:/TreeTools on your machine:

  oldWD <- getwd() # Remember the current working directory
  setwd(system.file(package = 'TreeTools'))

  # If taxon names were saved within the file (using `taxname=` in TNT),
  # then our job is easy:
  ReadTntTree('extdata/output/named.tre')

  # But if taxa were compressed to numbers (using `taxname-`), we need to
  # look up the original matrix in order to dereference the tip names.
  #
  # We need to extract the relevant file path from the end of the
  # hard-coded path in the original file.
  #
  # We are interested in the last two elements of
  # c:/TreeTools/input/dataset.nex
  #                2      1
  #
  # '.' means "relative to the current directory"
  ReadTntTree('extdata/output/numbered.tre', './extdata', 2)

  # If working in a lower subdirectory
  setwd('./extdata/otherfolder')

  # then it will be necessary to navigate up the directory path with '..':
  ReadTntTree('../output/numbered.tre', '..', 2)


  setwd(oldWD) # Restore original working directory

}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})

Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
