% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_numbering.R
\name{RenumberTree}
\alias{RenumberTree}
\alias{RenumberEdges}
\title{Renumber a tree in preorder}
\usage{
RenumberTree(parent, child)

RenumberEdges(parent, child, ...)
}
\arguments{
\item{parent}{Integer vector corresponding to the first column of the edge
matrix of a tree of class \code{\link{phylo}}, i.e. \code{tree$edge[, 1]}}

\item{child}{Integer vector corresponding to the second column of the edge
matrix of a tree of class \code{\link{phylo}}, i.e. \code{tree$edge[, 2]}.}

\item{\dots}{Ignored; included for compatibility with previous versions.}
}
\value{
\code{RenumberTree} returns an edge matrix for a tree of class \code{phylo}
following the usual preorder convention for edge and node numbering.

\code{RenumberEdges} formats the output of \code{RenumberTree()} into a list
whose two entries correspond to the new parent and child vectors.
}
\description{
Wrapper for the C function \code{preorder_edges_and_nodes}, which
renumbers internal nodes and orders edges in preorder, in an order
guaranteed to be identical for any tree of an equivalent topology.
At each node, child edges are arranged from left to right according to the
lowest-numbered tip in the subtree subtended by each edge.
}
\seealso{
Rotate each node into a consistent orientation with \code{\link[=SortTree]{SortTree()}}.

Other tree manipulation: 
\code{\link{AddTip}()},
\code{\link{CollapseNode}()},
\code{\link{ConsensusWithout}()},
\code{\link{DropTip}()},
\code{\link{EnforceOutgroup}()},
\code{\link{LeafLabelInterchange}()},
\code{\link{RenumberTips}()},
\code{\link{Renumber}()},
\code{\link{RootTree}()},
\code{\link{SingleTaxonTree}()},
\code{\link{SortTree}()},
\code{\link{Subtree}()}

Other C wrappers: 
\code{\link{Neworder}},
\code{\link{Reorder}}
}
\concept{C wrappers}
\concept{tree manipulation}
