% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reweight.R
\name{Reweight}
\alias{Reweight}
\title{Reweight phylogenetic characters
Reweight characters in a phylogenetic dataset}
\usage{
Reweight(dataset, weights)
}
\arguments{
\item{dataset}{A phylogenetic data matrix of \pkg{phangorn} class \code{phyDat},
or as a matrix in the format produced by \code{\link[=PhyDatToMatrix]{PhyDatToMatrix()}}.}

\item{weights}{Unnamed integer vector specifying desired weight of each
character in turn; or named integer vector specifying weights of each
character; unnamed entries will be assigned weight 1.}
}
\value{
\code{Reweight()} returns \code{dataset} after reweighting characters.
For a matrix, this will be attained by repeating each column the specified
number of times. For a \code{phyDat} object, the \code{weights} attribute will be
modified.
}
\description{
Reweight phylogenetic characters
Reweight characters in a phylogenetic dataset
}
\examples{
mat <- rbind(a = c(0, 2, 0), b = c(0, 2, 0), c = c(1, 3, 0), d = c(1, 3, 0))
dat <- MatrixToPhyDat(mat)

 # Set character 1 to weight 1, character 2 to weight 2; omit character 3
Reweight(mat, c(1, 2, 0))
# Equivalently:
Reweight(dat, c("3" = 0, "2" = 2))
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
