% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_files.R
\name{RightmostCharacter}
\alias{RightmostCharacter}
\title{Rightmost character of string}
\usage{
RightmostCharacter(string, len = nchar(string))
}
\arguments{
\item{string}{Character string.}

\item{len}{(Optional) Integer specifying number of characters in \code{string}.}
}
\value{
\code{RightmostCharacter()} returns the rightmost character of a string.
}
\description{
\code{RightmostCharacter()} is a convenience function that returns the final
character of a string.
}
\examples{
RightmostCharacter("Hello, World!")

}
\seealso{
Other string parsing functions: 
\code{\link{EndSentence}()},
\code{\link{MatchStrings}()},
\code{\link{MorphoBankDecode}()},
\code{\link{Unquote}()}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{string parsing functions}
