% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports-manual.R
\name{root_on_node}
\alias{root_on_node}
\title{Wrapper for internal C function \code{root_on_node()}}
\usage{
root_on_node(phy, outgroup)
}
\arguments{
\item{phy}{Minimally, a named list with entries \code{edge} and \code{Nnode}, in the
format of equivalent entries in a tree of class \code{phylo}. \code{edge.length} will
also be considered if supplied. The root node must be numbered \code{n_tip + 1},
per the \code{phylo} convention.}

\item{outgroup}{Integer specifying index of leaf or node to set as the
outgroup.}
}
\value{
\code{root_on_node()} returns \code{phy} rooted on the specified node, in
preorder.
}
\description{
Direct entry point to \code{root_on_node()}; recommended for expert use only.
\code{RootTree()} checks that input is properly formatted and is recommended
for general use.
}
\details{
\code{phy} may be supplied in any valid edge order; it is preordered internally,
so there is no need to call \code{\link[=Preorder]{Preorder()}} first.  The returned tree is
always in preorder, with node numbers reassigned accordingly.
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\keyword{internal}
