% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_shape.R
\name{RootedTreeShape}
\alias{RootedTreeShape}
\title{Integer representing shape of a tree}
\usage{
RootedTreeShape(tree)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}.}
}
\value{
\code{TreeShape} returns an integer specifying the shape of a tree,
ignoring tip labels.
}
\description{
Integer representing shape of a tree
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
