% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_rearrangement.R
\name{RootTree}
\alias{RootTree}
\title{Root a phylogenetic tree}
\usage{
RootTree(tree, outgroupTips)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}.}

\item{outgroupTips}{Character vector specifying the names of the tips to
include in the outgroup.}
}
\value{
\code{RootTree} returns a tree of class \code{phylo}, rooted on the smallest
clade that contains the specified tips.
}
\description{
Roots a tree on the smallest clade containing the specified tips.
}
\seealso{
\itemize{
\item \link[ape:root]{ape::root}
\item \code{\link{EnforceOutgroup}}
}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
