% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{SampleOne}
\alias{SampleOne}
\title{Quick sample}
\usage{
SampleOne(x, len = length(x))
}
\arguments{
\item{x}{vector to sample}

\item{len}{length of vector}
}
\value{
Returns a single, randomly sampled element of \code{x}.
}
\description{
Faster than inbuilt sample because it avoids some checks
}
\examples{
SampleOne(1:5, 5)

}
\keyword{internal}
