% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{SpectrumLegend}
\alias{SpectrumLegend}
\title{Spectrum legend
Deprecated wrapper for standalone function in package "PlotTools".}
\usage{
SpectrumLegend(...)
}
\arguments{
\item{\dots}{Parameters to pass to \code{PlotTools::SpectrumLegend()}.}
}
\description{
Spectrum legend
Deprecated wrapper for standalone function in package "PlotTools".
}
