% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsistentSplits.R
\name{SplitConsistent}
\alias{SplitConsistent}
\title{Are splits consistent?}
\usage{
SplitConsistent(needle, haystack)
}
\arguments{
\item{needle}{Splits object containing the single split to evaluate}

\item{haystack}{Splits object, or list thereof, containing the splits to
compare against \code{needle}.}
}
\value{
\code{SplitConsistent()} returns a list of logical vectors specifying,
for each entry in \code{haystack}, whether each split is consistent with \code{needle}.
}
\description{
Are splits consistent?
}
\examples{
splits1 <- as.Splits(BalancedTree(8))
splits2 <- as.Splits(PectinateTree(8))
summary(splits1[[4]])
SplitConsistent(splits1[[4]], splits2)
}
\seealso{
Other split manipulation functions: 
\code{\link{DropTip}()},
\code{\link{Subsplit}()},
\code{\link{TrivialSplits}()}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{split manipulation functions}
