% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stemwardness.R
\name{Stemwardness}
\alias{Stemwardness}
\title{'Stemwardness' of a leaf}
\usage{
Stemwardness(tree, tip)
}
\arguments{
\item{tree}{A tree of class \code{\link[ape:read.tree]{phylo}}.}

\item{tip}{Either a numeric specifying the index of a single tip, or a
character specifying its label.}
}
\value{
\code{Stemwardness()} returns a named vector of length two.
Entry \code{sisterSize} is an integer specifying the number of leaves in the clade
that is sister to \code{label}.
Entry \code{rootNodeDist} is an integer specifying the number of edges between
'label' and the root node.
}
\description{
\code{Stemwardness()} reports two measures of the position of a leaf relative
to the root.
}
\references{
\insertRef{Asher2020}{TreeTools}
}
\seealso{
Other tree characterization functions: 
\code{\link{CladisticInfo}()},
\code{\link{TotalCopheneticIndex}()}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree characterization functions}
