% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Support.R
\name{SupportColour}
\alias{SupportColour}
\alias{SupportColor}
\title{Colour for node support value}
\usage{
SupportColour(
  support,
  show1 = TRUE,
  scale = rev(diverge_hcl(101, h = c(260, 0), c = 100, l = c(50, 90), power = 1)),
  outOfRange = "red"
)

SupportColor(
  support,
  show1 = TRUE,
  scale = rev(diverge_hcl(101, h = c(260, 0), c = 100, l = c(50, 90), power = 1)),
  outOfRange = "red"
)
}
\arguments{
\item{support}{A numeric vector of values in the range 0--1.}

\item{show1}{Logical specifying whether to display values of 1.
A transparent white will be returned if \code{FALSE}.}

\item{scale}{101-element vector listing colours in sequence. Defaults to
a diverging HCL scale.}

\item{outOfRange}{Colour to use if results are outside the range 0--1.}
}
\value{
\code{SupportColour()} returns the appropriate value from \code{scale},
or \code{outOfRange} if a value is outwith the valid range.
}
\description{
Colour value with which to display node support.
}
\examples{
SupportColour((-1):4 / 4, show1 = FALSE)
}
