% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitFunctions.R
\name{TipLabels.numeric}
\alias{TipLabels.numeric}
\title{Extract tip names}
\usage{
\method{TipLabels}{numeric}(x)
}
\arguments{
\item{tips}{An integer specifying the number of tips, or a character vector
naming the tips, or a tree of class phylo, or a phylogenetic dataset whose
names will be used to name the tips.}
}
\value{
Character vector listing tip names.
}
\description{
Extracts tip labels from a dataset whose class may be one of \code{phyDat},
\code{Splits}, \code{phylo} or \code{list}.
}
\details{
If \code{tips} is numeric, will return a vector \code{t1}, \code{t2} ... \code{tn}, to match
the default of \link{ape:rtree}.
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\keyword{internal}
