% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopologyOnly.R
\name{TopologyOnly}
\alias{TopologyOnly}
\title{Remove metadata from trees}
\usage{
TopologyOnly(tree)
}
\arguments{
\item{tree}{A tree of class \code{\link[ape:read.tree]{phylo}}.}
}
\value{
Returns \code{tree}, with each tree in \code{\link{Preorder}}, with edge lengths,
node labels and other attributes removed.
}
\description{
\code{TopologyOnly()} removes all information from trees except for their
topologies and leaf labels.
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
