% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_information.R
\name{TreeInformation}
\alias{TreeInformation}
\alias{TreePhylogeneticInfo}
\alias{TreePhylogeneticInformation}
\title{Tree information content}
\usage{
TreePhylogeneticInfo(tree)

TreePhylogeneticInformation(tree)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}.}
}
\description{
Calculate the phylogenetic or clustering information content of a
phylogenetic tree.
}
\details{
#TODO
Detailed explanation to follow.
}
\seealso{
Other tree information functions: 
\code{\link{TreesMatchingTree}()}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree information functions}
