% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_properties.R
\name{TreeIsRooted}
\alias{TreeIsRooted}
\title{Is tree rooted?}
\usage{
TreeIsRooted(tree)
}
\arguments{
\item{tree}{A phylogenetic tree of class \code{phylo}.}
}
\value{
\code{TreeIsRooted()} returns a logical specifying whether a root node is
resolved.
}
\description{
\code{TreeIsRooted()} is a fast alternative to \code{ape::is.rooted()}.
}
\examples{
TreeIsRooted(BalancedTree(6))
TreeIsRooted(UnrootTree(BalancedTree(6)))

}
\seealso{
Other tree properties:
\code{\link[=Cherries]{Cherries()}},
\code{\link[=ConsensusWithout]{ConsensusWithout()}},
\code{\link[=EdgeRatio]{EdgeRatio()}},
\code{\link[=LongBranch]{LongBranch()}},
\code{\link[=MatchEdges]{MatchEdges()}},
\code{\link[=NSplits]{NSplits()}},
\code{\link[=NTip]{NTip()}},
\code{\link[=NodeNumbers]{NodeNumbers()}},
\code{\link[=PathLengths]{PathLengths()}},
\code{\link[=SplitsInBinaryTree]{SplitsInBinaryTree()}},
\code{\link[=TipLabels]{TipLabels()}},
\code{\link[=Treeness]{Treeness()}}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree properties}
