% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeTools-package.R
\docType{package}
\name{TreeTools-package}
\alias{TreeTools}
\alias{TreeTools-package}
\title{TreeTools: Create, Modify and Analyse Phylogenetic Trees}
\description{
Efficient implementations of functions for the creation, 
  modification and analysis of phylogenetic trees.
  Applications include: generation of trees with specified shapes;
  rooting of trees and extraction of subtrees;
  calculation and depiction of node support;
  calculation of ancestor-descendant relationships;
  import and export of trees from Newick, Nexus (Maddison et al. 1997)
  <doi:10.1093/sysbio/46.4.590>,
  and TNT <http://www.lillo.org.ar/phylogeny/tnt/> formats;
  and
  analysis of partitions and partition information.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ms609.github.io/TreeTools}
  \item \url{https://github.com/ms609/TreeTools}
  \item Report bugs at \url{https://github.com/ms609/TreeTools/issues}
}

}
\author{
\strong{Maintainer}: Martin R. Smith \email{martin.smith@durham.ac.uk} (\href{https://orcid.org/0000-0001-5660-1727}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Emmanuel Paradis (\href{https://orcid.org/0000-0003-3092-2199}{ORCID}) [copyright holder]
  \item Klaus Schliep (\href{https://orcid.org/0000-0003-2941-0161}{ORCID}) [copyright holder]
}

}
\keyword{internal}
