% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitFunctions.R
\name{TrivialSplits}
\alias{TrivialSplits}
\title{Are splits trivial?}
\usage{
TrivialSplits(splits, nTip = attr(splits, "nTip"))
}
\arguments{
\item{splitsObjectParam}{}
}
\value{
Logical vector specifying whether each split in \code{splits} is trivial,
i.e. includes or excludes only a single tip or no tips at all.
}
\description{
Are splits trivial?
}
\seealso{
Other split manipulation functions: \code{\link{Subsplit}}
}
\author{
Martin R. Smith
}
\concept{split manipulation functions}
