% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_files.R
\name{Unquote}
\alias{Unquote}
\title{Remove quotation marks from a string}
\usage{
Unquote(string)
}
\arguments{
\item{string}{Input string}
}
\value{
\code{Unquote()} returns \code{string}, with any matched punctuation marks
and trailing whitespace removed.
}
\description{
Remove quotation marks from a string
}
\examples{
Unquote("'Hello World'")
}
\seealso{
Other string parsing functions: 
\code{\link{EndSentence}()},
\code{\link{MatchStrings}()},
\code{\link{MorphoBankDecode}()},
\code{\link{RightmostCharacter}()}
}
\author{
Martin R. Smith
}
\concept{string parsing functions}
