% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Splits.R
\name{xor}
\alias{xor}
\alias{xor,Splits,Splits-method}
\title{Exclusive OR operation}
\usage{
xor(x, y)

\S4method{xor}{Splits,Splits}(x, y)
}
\arguments{
\item{x, y}{Objects to be compared.}
}
\description{
Exclusive OR operation
}
\seealso{
Other Splits operations:
\code{\link[=LabelSplits]{LabelSplits()}},
\code{\link[=NSplits]{NSplits()}},
\code{\link[=NTip]{NTip()}},
\code{\link[=PolarizeSplits]{PolarizeSplits()}},
\code{\link[=SplitFrequency]{SplitFrequency()}},
\code{\link{Splits}},
\code{\link[=SplitsInBinaryTree]{SplitsInBinaryTree()}},
\code{\link[=TipLabels]{TipLabels()}},
\code{\link[=TipsInSplits]{TipsInSplits()}},
\code{\link{match,Splits,Splits-method}}
}
\concept{Splits operations}
