% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Splits.R
\name{as.Splits}
\alias{as.Splits}
\alias{as.Splits.phylo}
\title{As splits}
\usage{
as.Splits(x, tipLabels = NULL, ...)

\method{as.Splits}{phylo}(x, tipLabels = NULL, asSplits = TRUE, ...)
}
\arguments{
\item{x}{Object to convert into splits: perhaps a tree of class
\code{\link[ape:read.tree]{phylo}}.}

\item{tipLabels}{Character vector specifying sequence in which to order
tip labels.  Label order must (currently) match to combine or compare separate
\code{Splits} objects.}

\item{\dots}{Presently unused.}

\item{asSplits}{Logical specifying whether to return a \code{Splits} object,
or an unannotated two-dimensional array (useful where performance is
paramount).}
}
\value{
Returns an object of class \code{Splits}, or (if \code{asSplits = FALSE}) a
two-dimensional array of 32-bit integers, which each bit specifying whether
a tip is a member of the split.  Splits are named according to the node
that defines them.
}
\description{
Converts a phylogenetic tree to an array of bipartition splits.
}
\section{Methods (by class)}{
\itemize{
\item \code{phylo}: Convert object of class \code{phylo} to \code{Splits}.
}}

\examples{
splits <- as.Splits(BalancedTree(letters[1:6]))
summary(splits)
TipsInSplits(splits)
summary(!splits)
TipsInSplits(!splits)

length(splits + !splits)
length(unique(splits + !splits))

summary(c(splits[[2:3]], !splits[[1:2]]))

moreSplits <- as.Splits(PectinateTree(letters[6:1]), tipLabel = splits)
print(moreSplits, details = TRUE)
match.Splits(splits, moreSplits)
in.Splits(moreSplits, splits)

}
\seealso{
Other Splits operations: \code{\link{NSplits}},
  \code{\link{TipsInSplits}}, \code{\link{in.Splits}},
  \code{\link{match.Splits}}
}
\author{
Martin R. Smith
}
\concept{Splits operations}
