% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirthDeath.R
\name{BirthDeath}
\alias{BirthDeath}
\title{Simulate trees under the birth-death process}
\usage{
BirthDeath()
}
\arguments{
\item{pi}{Vector of probabilities associated with each of \emph{d} types}

\item{lambda}{Square numeric matrix giving rate at which a lineage of type
\emph{a} gives birth (via cladogenesis) to an additional lineage of type \emph{b}.
Entries on the diagonal give rate of (cladogenic) birth without mutation.}

\item{mu}{Numeric vector giving death rate for a lineage of each type}

\item{psi}{Numeric vector giving sampling rate for a lineage of each type}

\item{rA}{Numeric vector giving probability that a lineage of each type
dies after sampling}

\item{gamma}{Rate of anagenic mutation (i.e. change of type without a birth).}

\item{rAL}{Numeric vector giving probability of dying after a concerted
sampling event.}

\item{tMax}{time at which to start simulation (in units of time ago)}

\item{nMax}{Abort simulation when more than \code{nMax} lineages exist

Uses the forward-equivalent simulation described by
\insertRef{Celentano2024}{TreeTools}
https://arxiv.org/pdf/2402.17153v1.pdf}
}
\description{
Simulate trees under the birth-death process
}
\references{
\insertAllCited{}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
