% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Information.R
\name{CharacterInformation}
\alias{CharacterInformation}
\title{Character information content}
\usage{
CharacterInformation(tokens)
}
\arguments{
\item{tokens}{Character vector specifying the tokens assigned to each taxon for
a character.  Example: \code{c(0, 0, 0, 1, 1, 1, '?', '-')}.

Note that ambiguous tokens such as \code{(01)} are not supported, and should be
replaced with \code{?}.`}
}
\value{
Phylogenetic information content of the character, in bits.
}
\description{
Calculates the phylogenetic information content of a given character.
}
\seealso{
Other split information functions: \code{\link{SplitInformation}},
  \code{\link{TreesMatchingSplit}},
  \code{\link{UnrootedTreesMatchingSplit}}
}
\author{
Martin R. Smith
}
\concept{split information functions}
