% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_information.R
\name{CharacterMI}
\alias{CharacterMI}
\alias{CharacterEntropy}
\alias{ApportionAmbiguity}
\alias{JointCharacterEntropy}
\title{Mutual information between characters}
\usage{
CharacterEntropy(char, ignore = character(0))

ApportionAmbiguity(char, ignore = character(0))

JointCharacterEntropy(char1, char2, ignore = character(0))
}
\arguments{
\item{char, char1, char2}{Character vector listing character tokens.
For handling of ambiguous tokens, see section 'Ambiguity'.}

\item{ignore}{Character vector listing tokens to treat as ambiguous,
besides \verb{?} and \code{-}.}
}
\description{
Mutual information between characters
}
\section{Ambiguity}{

Ambiguous characters should be specified as a string containing all possible
characters states (and, optionally, parentheses): e.g. \verb{[01]}.
\verb{?} and \code{-} will be interpreted as 'any possible value'.
Each token is assigned a probability according to its number of
\emph{non-ambiguous} occurrences in the character.
Example: in \verb{000000 1111 [01] [02] ?}, \verb{[01]} and \verb{?} will be taken to have
a 6/10 chance of being token \code{0};
\verb{[02]} will be treated as having a 100\% chance of being state \code{0}.
}

\examples{
data('Lobo', package = 'TreeTools')
dat <- PhyDatToMatrix(Lobo.phy)

characterEntropies <- apply(dat, 2, CharacterEntropy)
head(characterEntropies)


char <- c(rep('0', 6), rep('1', 4), '[01]', '[02]', '?')
ShareAmbiguity(char)
ShareAmbiguity(char, ignore = '[02]')
CharacterEntropy(char)
}
\references{
\insertRef{Dunn2008}{TreeTools}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
