% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cherries.R
\name{Cherries}
\alias{Cherries}
\alias{Cherries.phylo}
\alias{Cherries.numeric}
\title{Number of cherries in a phylogenetic tree}
\usage{
Cherries(tree, nTip)

\method{Cherries}{phylo}(tree, nTip = NTip(tree))

\method{Cherries}{numeric}(tree, nTip)
}
\arguments{
\item{tree}{A binary phylogenetic tree, of class \code{phylo}; or a matrix
corresponding to its edge matrix.}

\item{nTip}{Number of leaves in tree.}
}
\value{
\code{Cherries()} returns an integer specifying the number of nodes whose
children are both leaves.
}
\description{
Number of cherries in a phylogenetic tree
}
\seealso{
Other tree properties: 
\code{\link{ConsensusWithout}()},
\code{\link{LongBranch}()},
\code{\link{MatchEdges}()},
\code{\link{NSplits}()},
\code{\link{NTip}()},
\code{\link{NodeNumbers}()},
\code{\link{PathLengths}()},
\code{\link{SplitsInBinaryTree}()},
\code{\link{TipLabels}()},
\code{\link{TreeIsRooted}()},
\code{\link{Treeness}()}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree properties}
