% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_information.R
\name{ClusteringInfo}
\alias{ClusteringInfo}
\title{Clustering Information}
\usage{
ClusteringInfo(x)
}
\arguments{
\item{x}{Tree of class \code{phylo}, or a list thereof.}
}
\description{
#TODO
}
\details{
update TreeDist docs to refer to this function.

Unbalanced trees contain more uneven clusters, and thus contain
more clustering information than balanced trees.

This is considered a defect by Rohlf (1982) and Page (1992)
}
\examples{
tree <- CollapseNode(BalancedTree(9), 13:15)
plot(tree)
nodelabels()
ClusteringInfo(tree)
ClusteringInfo(BalancedTree(8))
ClusteringInfo(PectinateTree(8))

}
\references{
\insertRef{Page1992}{TreeTools}

\insertRef{Rohlf1982}{TreeTools}
}
\seealso{
Other tree information functions: 
\code{\link{PhylogeneticInfo}()},
\code{\link{TreesMatchingTree}()}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree information functions}
