% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitFunctions.R
\name{CompatibleSplits}
\alias{CompatibleSplits}
\title{Which splits are compatible?}
\usage{
CompatibleSplits(splits, splits2)
}
\arguments{
\item{splits2}{A second \code{Splits} object.}

\item{split}{An object of class \link{as.Splits}{`Splits`}.}
}
\value{
A logical matrix specifying whether each split in \code{splits} is
compatible with each split in \code{splits2}.
}
\description{
Which splits are compatible?
}
\examples{

splits <- as.Splits(BalancedTree(8))
splits2 <- as.Splits(PectinateTree(8))

summary(splits)
summary(splits2)

CompatibleSplits(splits, splits2)

}
\author{
Martin R. Smith
}
