% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_properties.R
\name{DescendantEdges}
\alias{DescendantEdges}
\alias{DescendantTips}
\alias{AllDescendantEdges}
\title{Identify descendant edges}
\usage{
DescendantEdges(parent, child, edge = NULL, nEdge = length(parent))

DescendantTips(
  parent,
  child,
  edge = NULL,
  nEdge = length(parent),
  nTip = min(parent) - 1L
)

AllDescendantEdges(parent, child, nEdge = length(parent))
}
\arguments{
\item{parent}{Integer vector corresponding to the first column of the edge
matrix of a tree of class \code{\link{phylo}}, i.e. \code{tree[["edge"]][, 1]}}

\item{child}{Integer vector corresponding to the second column of the edge
matrix of a tree of class \code{\link{phylo}}, i.e. \code{tree[["edge"]][, 2]}.}

\item{edge}{Integer specifying the number of the edge whose children are
required (see \code{\link[ape:nodelabels]{edgelabels}()}).}

\item{nEdge}{number of edges (calculated from \code{length(parent)} if not
supplied).}

\item{nTip}{Integer specifying number of leaves in the tree.
Computed if not supplied.}
}
\value{
\code{DescendantEdges()} returns a logical vector stating whether each
edge in turn is the specified edge or one of its descendants.

\code{DescendantTips()} returns a logical vector stating whether each
leaf in turn is a descendant of the specified edge.

\code{AllDescendantEdges()} is deprecated; use \code{DescendantEdges()}
instead.
It returns a matrix of class logical, with row \emph{N} specifying whether each
edge is a descendant of edge \emph{N} (or the edge itself).
}
\description{
\code{DescendantEdges()} efficiently identifies edges that are "descended" from
edges in a tree.

\code{DescendantTips()} efficiently identifies leaves (external nodes) that are
"descended" from edges in a tree.
}
\examples{
tree <- as.phylo(0, 6)
plot(tree)
desc <- DescendantEdges(tree$edge[, 1], tree$edge[, 2], edge = 5)
which(desc)
ape::edgelabels(bg = 3 + desc)
tips <- DescendantTips(tree$edge[, 1], tree$edge[, 2], edge = 5)
which(tips)
tiplabels(bg = 3 + tips)
}
\seealso{
Other tree navigation: 
\code{\link{AncestorEdge}()},
\code{\link{CladeSizes}()},
\code{\link{EdgeAncestry}()},
\code{\link{EdgeDistances}()},
\code{\link{ListAncestors}()},
\code{\link{MRCA}()},
\code{\link{NDescendants}()},
\code{\link{NodeDepth}()},
\code{\link{NodeNumbers}()},
\code{\link{NodeOrder}()},
\code{\link{RootNode}()}
}
\concept{tree navigation}
