% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_properties.R
\name{EdgeDistances}
\alias{EdgeDistances}
\title{Distance between edges}
\usage{
EdgeDistances(tree)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}.}
}
\description{
Number of nodes that must be traversed to navigate from each edge to
each other edge within a tree
}
\seealso{
Other tree navigation: \code{\link{AllAncestors}},
  \code{\link{AncestorEdge}},
  \code{\link{DescendantEdges}},
  \code{\link{EdgeAncestry}}, \code{\link{MRCA}},
  \code{\link{NonDuplicateRoot}}
}
\author{
Martin R. Smith
}
\concept{tree navigation}
