% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EdgeRatio.R
\name{EdgeRatio}
\alias{EdgeRatio}
\alias{EdgeRatio.phylo}
\title{Ratio of external:internal edge length}
\usage{
EdgeRatio(x)

\method{EdgeRatio}{phylo}(x)
}
\arguments{
\item{x}{A tree of class \code{\link[ape:read.tree]{phylo}}.}
}
\value{
\code{EdgeRatio()} returns a numeric specifying the ratio of external
to internal edge length (> 1 means the length of a tree is predominantly
in external edges), with attributes \code{external}, \code{internal}, and \code{total}
specifying the total length associated with edges of that nature.
}
\description{
Reports the ratio of tree length associated with external edges (i.e.
edges whose child is a leaf) and internal edges.
Where tree length is dominated by internal edges, variation between tips
is dominantly controlled by phylogenetic history.
}
\seealso{
Other tree properties: 
\code{\link{Cherries}()},
\code{\link{ConsensusWithout}()},
\code{\link{LongBranch}()},
\code{\link{MatchEdges}()},
\code{\link{NSplits}()},
\code{\link{NTip}()},
\code{\link{NodeNumbers}()},
\code{\link{PathLengths}()},
\code{\link{SplitsInBinaryTree}()},
\code{\link{TipLabels}()},
\code{\link{TreeIsRooted}()},
\code{\link{Treeness}()}
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree properties}
