% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_files.R
\name{ExtractTaxa}
\alias{ExtractTaxa}
\alias{NexusTokens}
\title{Extract taxa from a matrix block}
\usage{
ExtractTaxa(matrixLines, character_num = NULL, session = NULL)

NexusTokens(tokens, character_num = NULL, session = NULL)
}
\arguments{
\item{matrixLines}{lines of a file containing a phylogenetic matrix
(see ReadCharacters for expected format)}

\item{character_num}{Index of character(s) to return.
\code{NULL}, the default, returns all characters.}

\item{session}{(optionally) a Shiny session with a numericInput named character_num
whose maximum should be updated.}

\item{tokens}{Vector of character strings correponding to phylogenetic
tokens.}
}
\value{
Matrix with n rows, each named for the relevant taxon, and c columns,
each corresponding to the respective character specified in \code{character_num}
}
\description{
Extract taxa from a matrix block
}
\section{Functions}{
\itemize{
\item \code{NexusTokens}: Extract tokens from a string
}}

\keyword{internal}
