% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeptNodes.R
\name{KeptVerts}
\alias{KeptVerts}
\alias{KeptVerts.phylo}
\alias{KeptVerts.numeric}
\title{Identify vertices retained when leaves are dropped}
\usage{
KeptVerts(tree, keptTips)

\method{KeptVerts}{phylo}(tree, keptTips)

\method{KeptVerts}{numeric}(tree, keptTips)
}
\arguments{
\item{tree}{Original tree of class \code{phylo}, in \code{\link{Preorder}}.}

\item{keptTips}{Logical vector stating whether each leaf should be retained.
Sequence corresponds to \code{tree[["tip.label"]]}.}
}
\description{
Identify vertices retained when leaves are dropped
}
\examples{
master <- BalancedTree(12)
master <- Preorder(master) # Nodes must be listed in Preorder sequence
plot(master)
nodelabels()

allTips <- master[["tip.label"]]
keptTips <- sample(allTips, 8)
plot(KeepTip(master, keptTips))
KeptVerts(master, allTips \%in\% keptTips)
}
\seealso{
Other tree manipulation: 
\code{\link{AddTip}()},
\code{\link{CollapseNode}()},
\code{\link{ConsensusWithout}()},
\code{\link{DropTip}()},
\code{\link{EnforceOutgroup}()},
\code{\link{ImposeConstraint}()},
\code{\link{KeptPaths}()},
\code{\link{LeafLabelInterchange}()},
\code{\link{MakeTreeBinary}()},
\code{\link{RenumberTips}()},
\code{\link{RenumberTree}()},
\code{\link{Renumber}()},
\code{\link{RootTree}()},
\code{\link{SingleTaxonTree}()},
\code{\link{SortTree}()},
\code{\link{Subtree}()}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree manipulation}
