% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lobo.data}
\alias{Lobo.data}
\alias{Lobo.phy}
\title{Data from Zhang et al. 2016}
\format{
An object of class \code{list} of length 48.

An object of class \code{phyDat} of length 48.
}
\source{
\insertRef{Zhang2016}{TreeTools}
}
\usage{
Lobo.data

Lobo.phy
}
\description{
Phylogenetic data from Zhang \emph{et al.} (2016) in raw (\code{Lobo.data}) and
\code{phyDat} (\code{Lobo.phy}) formats.
}
\keyword{datasets}
