% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_files.R
\name{MatrixToPhyDat}
\alias{MatrixToPhyDat}
\alias{PhyDatToMatrix}
\title{Matrix to \code{phyDat}}
\usage{
MatrixToPhyDat(tokens)

PhyDatToMatrix(dataset)
}
\arguments{
\item{tokens}{matrix of tokens, probably created with \code{\link{ReadCharacters}}
or \code{\link{ReadTntCharacters}}. Row names should correspond to tip
labels; column names may optionally correspond to
character labels.}

\item{dataset}{A dataset of class \code{phyDat}.}
}
\value{
an object of class \code{phyDat}.

\code{PhyDatToMatrix()} returns a matrix corresponding to the
uncompressed character states within a \code{phyDat} object.
}
\description{
Converts a matrix of tokens to a \code{phyDat} object.
}
\section{Functions}{
\itemize{
\item \code{PhyDatToMatrix}: Converts a \code{phyDat} object to a matrix of tokens.
}}

\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\keyword{internal}
