% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_files.R
\name{MatrixToPhyDat}
\alias{MatrixToPhyDat}
\alias{PhyDatToMatrix}
\title{Matrix to phydat}
\usage{
MatrixToPhyDat(tokens)

PhyDatToMatrix(dataset)
}
\arguments{
\item{tokens}{matrix of tokens, probably created with \link{ReadCharacters}
or \link{ReadTntCharacters}. Row names should correspond to tip
labels; column names may optionally correspond to
character labels.}

\item{dataset}{A dataset of class \code{phyDat}.}
}
\value{
an object of class \code{phyDat}

A matrix corresponding to the uncompressed character states within
a phyDat object.
}
\description{
Converts a matrix of tokens to a phyDat object
}
\section{Functions}{
\itemize{
\item \code{PhyDatToMatrix}: Converts a phyDat object to a matrix of tokens.
}}

\author{
Martin R. Smith
}
\keyword{internal}
