% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_files.R
\name{MorphoBankDecode}
\alias{MorphoBankDecode}
\title{Decode MorphoBank text}
\usage{
MorphoBankDecode(string)
}
\arguments{
\item{string}{String to process}
}
\value{
\code{MorphoBankDecode()} returns a string with new lines and punctuation
reformatted.
}
\description{
Converts strings from MorphoBank notes into a Latex-compatible format.
}
\seealso{
Other string parsing functions: 
\code{\link{EndSentence}()},
\code{\link{MatchStrings}()},
\code{\link{RightmostCharacter}()},
\code{\link{Unquote}()}
}
\author{
Martin R. Smith
}
\concept{string parsing functions}
