% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_properties.R
\name{MRCA}
\alias{MRCA}
\title{Most Recent Common Ancestor}
\usage{
MRCA(tip1, tip2, ancestors)
}
\arguments{
\item{tip1, tip2}{Integer specifying index of tips whose most recent common
ancestor should be found.}

\item{ancestors}{Output of \code{\link{AllAncestors}} for the tree in question}
}
\description{
Most Recent Common Ancestor
}
\seealso{
Other tree navigation: \code{\link{AllAncestors}},
  \code{\link{AncestorEdge}},
  \code{\link{DescendantEdges}},
  \code{\link{EdgeAncestry}},
  \code{\link{NonDuplicateRoot}}
}
\author{
Martin R. Smith
}
\concept{tree navigation}
