% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combinatorics.R
\name{N1Spr}
\alias{N1Spr}
\alias{IC1Spr}
\title{Number of trees one SPR step away}
\usage{
N1Spr(n)

IC1Spr(n)
}
\arguments{
\item{n}{Integer vector specifying the number of tips in a tree.}
}
\value{
\code{N1Spr()} returns an integer vector denoting the number of trees one
SPR rearrangement away from the input tree..

\code{IC1Spr()} returns an numeric vector giving the phylogenetic
information content of trees 0 or 1 SPR rearrangement from an \emph{n}-leaf tree,
in bits.
}
\description{
\code{N1Spr()} calculates the number of trees one subtree prune-and-regraft
operation away from a binary input tree using the formula given by
\insertCite{Allen2001;textual}{BigTreeTools};
\code{IC1Spr()} calculates the information content of trees at this
distance: i.e. the entropy corresponding to the proportion of all possible
\emph{n}-tip trees whose SPR distance is at most one from a specified tree.
}
\examples{
N1Spr(4:6)
IC1Spr(5)

}
\references{
\insertAllCited{}
}
