% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_properties.R
\name{NDescendants}
\alias{NDescendants}
\title{Number of descendants for each node in a tree}
\usage{
NDescendants(tree)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}.}
}
\value{
\code{NDescendants()} returns a table listing the number of direct
descendants for each node in a tree.
}
\description{
Number of descendants for each node in a tree
}
\examples{
tree <- CollapseNode(BalancedTree(8), 12:15)
plot(tree)
nodelabels()
NDescendants(tree)

}
\seealso{
Other tree navigation: 
\code{\link{AllAncestors}()},
\code{\link{AncestorEdge}()},
\code{\link{DescendantEdges}()},
\code{\link{EdgeAncestry}()},
\code{\link{EdgeDistances}()},
\code{\link{MRCA}()},
\code{\link{NodeOrder}()},
\code{\link{NonDuplicateRoot}()},
\code{\link{RootNode}()}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree navigation}
