% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_numbering.R
\name{NeworderPruningwise}
\alias{NeworderPruningwise}
\title{neworder_pruningwise}
\usage{
NeworderPruningwise(nTip, nb.node, parent, child, nb.edge)
}
\description{
Wrapper for the ape function
}
\seealso{
Other C wrappers: \code{\link{Cladewise}},
  \code{\link{NeworderPhylo}}, \code{\link{RenumberTree}}
}
\concept{C wrappers}
\keyword{internal}
