% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combinatorics.R
\name{NRooted}
\alias{NRooted}
\alias{NUnrooted}
\alias{LnUnrooted}
\alias{LnUnrooted.int}
\alias{LnRooted}
\alias{LnRooted.int}
\alias{LnUnrootedSplits}
\alias{NUnrootedSplits}
\alias{LnUnrootedMult}
\alias{NUnrootedMult}
\title{Number of rooted/unrooted trees}
\usage{
NRooted(tips)

NUnrooted(tips)

LnUnrooted(tips)

LnUnrooted.int(tips)

LnRooted(tips)

LnRooted.int(tips)

LnUnrootedSplits(splits)

NUnrootedSplits(splits)

LnUnrootedMult(splits)

NUnrootedMult(splits)
}
\arguments{
\item{tips}{Integer specifying the number of tips.}

\item{splits}{Integer vector listing the number of taxa in each tree
bipartition.}
}
\description{
These functions return the number of rooted or unrooted trees consistent with
a given pattern of splits.
}
\details{
Functions starting N return the number of rooted or unrooted trees, functions
starting Ln provide the natural logarithm of this number.
Calculations follow Carter \emph{et al.} 1990, Theorem 2.
}
\section{Functions}{
\itemize{
\item \code{NUnrooted}: Number of unrooted trees

\item \code{LnUnrooted}: Log Number of unrooted trees

\item \code{LnUnrooted.int}: Log Number of unrooted trees (as integer)

\item \code{LnRooted}: Log Number of rooted trees

\item \code{LnRooted.int}: Log Number of rooted trees (as integer)

\item \code{LnUnrootedSplits}: Log number of unrooted trees

\item \code{NUnrootedSplits}: Number of unrooted trees consistent with a bipartition
split.

\item \code{LnUnrootedMult}: Log unrooted with multi-partition splits.

\item \code{NUnrootedMult}: Number of unrooted trees consistent with a multi-partition
split.
}}

\examples{
  NRooted(10)
  NUnrooted(10)
  LnRooted(10)
  LnUnrooted(10)
  # Number of trees consistent with a character whose states are
  # 00000 11111 222
  NUnrootedMult(c(5,5,3))

LnUnrootedSplits(c(2,4))
LnUnrootedSplits(c(3,3))
NUnrootedSplits(c(2,4))
NUnrootedSplits(c(3,3))
}
\references{
\insertRef{Carter1990}{TreeTools}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{split information function}
