% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nRootedShapes}
\alias{nRootedShapes}
\alias{nUnrootedShapes}
\title{Number of rooted / unrooted tree shapes}
\format{
An object of class \code{integer64} of length 55.

An object of class \code{integer64} of length 60.
}
\source{
\code{nRootedShapes} corresponds to the Wedderburn-Etherington numbers,
\href{https://oeis.org/A001190}{\acronym{OEIS} A001190}

\code{nUnrootedShapes} is \href{https://oeis.org/A000672}{\acronym{OEIS} A000672}
}
\usage{
nRootedShapes

nUnrootedShapes
}
\description{
\code{nRootedShapes} and \code{nUnrootedShapes} give the number of (un)rooted binary
trees on \emph{n} unlabelled leaves.
}
\keyword{datasets}
