% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_information.R
\name{PhylogeneticInfo}
\alias{PhylogeneticInfo}
\alias{PhylogeneticInformation}
\title{Phylogenetic information content}
\usage{
PhylogeneticInfo(x)

PhylogeneticInformation(x)
}
\arguments{
\item{x}{Tree of class \code{phylo}, or a list thereof.}
}
\value{
Returns the phylogenetic or clustering information content
of the input tree(s).
}
\description{
Calculate the phylogenetic information content of a phylogenetic object.
}
\details{
#TODO
Splits to follow
}
\seealso{
Clustering information content: coming soon. #TODO.

Other tree information functions: 
\code{\link{TreesMatchingTree}()}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree information functions}
