% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Splits.R
\name{PolarizeSplits}
\alias{PolarizeSplits}
\title{Polarize splits on a single taxon}
\usage{
PolarizeSplits(x, pole = 1L)
}
\arguments{
\item{x}{Object of class \code{\link{Splits}}.}

\item{pole}{Numeric or character identifying tip that should polarize each
split.}
}
\value{
\code{PolarizeSplits()} returns a \code{Splits} object in which \code{pole} is
represented by a zero bit
}
\description{
Polarize splits on a single taxon
}
\seealso{
Other Splits operations: 
\code{\link{LabelSplits}()},
\code{\link{NSplits}()},
\code{\link{NTip}()},
\code{\link{SplitFrequency}()},
\code{\link{SplitsInBinaryTree}()},
\code{\link{Splits}},
\code{\link{TipLabels}()},
\code{\link{TipsInSplits}()},
\code{\link{match.Splits}},
\code{\link{xor}()}
}
\concept{Splits operations}
