% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_numbering.R
\name{RenumberTree}
\alias{RenumberTree}
\alias{RenumberEdges}
\title{Renumber a tree}
\usage{
RenumberTree(parent, child)

RenumberEdges(parent, child, nEdge = length(parent))
}
\value{
an edge matrix for a tree in following the usual preorder convention
for edge and node numbering
}
\description{
Order edges and number nodes
}
\details{
Wrapper for the C function \code{preorder_edges_and_nodes}
}
\section{Functions}{
\itemize{
\item \code{RenumberEdges}: Instead returns a list containing two items
corresponding to the new parent and child vectors
}}

\seealso{
Other C wrappers: \code{\link{Cladewise}},
  \code{\link{NeworderPhylo}},
  \code{\link{NeworderPruningwise}}

Other C wrappers: \code{\link{Cladewise}},
  \code{\link{NeworderPhylo}},
  \code{\link{NeworderPruningwise}}
}
\concept{C wrappers}
\keyword{internal}
