% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_files.R
\name{RightmostCharacter}
\alias{RightmostCharacter}
\title{Rightmost character of string}
\usage{
RightmostCharacter(string, len = nchar(string))
}
\arguments{
\item{string}{Character string.}

\item{len}{Integer specifying number of characters in \code{string}.}
}
\value{
Rightmost character of \code{string}.
}
\description{
Rightmost character of string
}
\examples{
RightmostCharacter("Hello, World!")

}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\keyword{internal}
