% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports-manual.R
\name{root_on_node}
\alias{root_on_node}
\title{C++ wrapper for root_on_node}
\usage{
root_on_node(phy, outgroup)
}
\arguments{
\item{phy}{Tree of class \code{phylo}. root node must == n_tip + 1.}

\item{outgroup}{Integer specifying new root node.}
}
\description{
Wrapper for C++ function \code{root_on_node()}.
No attempt is made to validate input; specifying invalid input may crash
R.
Behaviour is subject to change without notice.
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
