% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports-manual.R
\name{root_on_node}
\alias{root_on_node}
\title{Wrapper for internal C function \code{root_on_node()}}
\usage{
root_on_node(phy, outgroup)
}
\arguments{
\item{phy}{Minimally, a named list with entries \code{edge} and \code{Nnode}, in the
format of equivalent entries in a tree of class \code{phylo}. \code{edge.length} will
also be considered if supplied.}

\item{outgroup}{Numeric specifying index of leaf to use as outgroup.}
}
\value{
\code{root_on_node()} returns \code{phy} rooted on the speficied leaf.
}
\description{
Direct entry point to \code{root_on_node()}; recommended for expert use only.
\code{RootTree()} checks that input is properly formatted and is recommended
for general use.
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\keyword{internal}
