% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_properties.R
\name{RootNode}
\alias{RootNode}
\title{Which node is a tree's root?}
\usage{
RootNode(tree)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}.}
}
\value{
\code{RootNode()} returns an integer denoting the root node.
}
\description{
Identify the root node of a (rooted or unrooted) phylogenetic tree.
Unrooted trees are represented internally by a rooted tree with an
unresolved root node.
}
\examples{
RootNode(BalancedTree(8))
RootNode(unroot(BalancedTree(8)))


}
\seealso{
\code{\link[=TreeIsRooted]{TreeIsRooted()}}

phangorn::\code{\link[=getRoot]{getRoot()}}

Other tree navigation: \code{\link{AllAncestors}},
  \code{\link{AncestorEdge}},
  \code{\link{DescendantEdges}},
  \code{\link{EdgeAncestry}}, \code{\link{EdgeDistances}},
  \code{\link{MRCA}}, \code{\link{NonDuplicateRoot}}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\concept{tree navigation}
