% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{SampleOne}
\alias{SampleOne}
\title{Quick sample}
\usage{
SampleOne(x, len = length(x))
}
\arguments{
\item{x}{A vector to sample.}

\item{len}{(optionally) integer specifying length of \code{x}.}
}
\value{
A vector of length one, randomly sampled from \code{x}.
}
\description{
Faster than inbuilt \code{\link{sample}} because it avoids some checks.
}
\examples{
SampleOne(9:10)
SampleOne(letters[1:4])

}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
\keyword{internal}
