% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort.R
\name{sort.multiPhylo}
\alias{sort.multiPhylo}
\alias{==.phylo}
\alias{<.phylo}
\alias{>.phylo}
\title{Sort a list of phylogenetic trees}
\usage{
\method{sort}{multiPhylo}(x, decreasing = FALSE, na.last = NA, ...)

\method{==}{phylo}(e1, e2)

\method{<}{phylo}(e1, e2)

\method{>}{phylo}(e1, e2)
}
\arguments{
\item{x, decreasing, na.last, \dots}{As in \code{\link[=sort]{sort()}}.}

\item{e1, e2}{\code{phylo} objects to be compared.}
}
\description{
Trees are sorted alphabetically by their Newick representation after
numbering tips alphabetically and renumbering tres in \code{\link{Preorder}}.
}
\details{
This method of sorting is inelegant, and will be replaced in a future
version by a method that is more principled
(\href{https://github.com/ms609/TreeTools/issues/84}{#84}).
}
\examples{
sort(as.phylo(0:5, 7))
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
