% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_display.R
\name{SortTree}
\alias{SortTree}
\title{Sort tree}
\usage{
SortTree(tree)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}.}
}
\value{
\code{SortTree} returns a tree of class \code{phylo}, with each node sorted
such that the larger clade is first.
}
\description{
Sorts each node into a consistent order, so similar trees look visually
similar.
}
\seealso{
\code{\link{RenumberTree}}
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
