% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{SpectrumLegend}
\alias{SpectrumLegend}
\title{Produce a legend for continuous gradient scales}
\usage{
SpectrumLegend(
  x0 = 0.05,
  y0 = 0.05,
  x1 = x0,
  y1 = y0 + 0.2,
  absolute = FALSE,
  legend,
  palette,
  ...
)
}
\arguments{
\item{x0, y0, x1, y1}{Coordinates of the bottom-left and top-right end of the
bar.}

\item{absolute}{Logical specifying whether \code{x0}, \code{y1}, etc denote
coordinates (\code{TRUE}) or relative position, where (0, 0) denotes the
bottom-left of the plot area and (1, 1) the top right.}

\item{legend}{Character vector with which to label points on \code{palette}.}

\item{palette}{Colour palette to depict.}

\item{\dots}{Additional parameters to \code{\link[=segments]{segments()}}.}

\item{xSize, ySize}{Length of bar, in absolute units or as a proportion of the
plotting area.}
}
\description{
Prints an annotated vertical bar coloured according to a continuous palette.
}
\details{
This convenience function is not yet very customizable; do file a GitHub
issue if you would value additional functionality.
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
