% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitFunctions.R
\name{TipLabels}
\alias{TipLabels}
\alias{TipLabels.phylo}
\alias{TipLabels.Splits}
\alias{TipLabels.list}
\alias{TipLabels.multiPhylo}
\alias{TipLabels.character}
\title{Tip labels}
\usage{
TipLabels(x)

\method{TipLabels}{phylo}(x)

\method{TipLabels}{Splits}(x)

\method{TipLabels}{list}(x)

\method{TipLabels}{multiPhylo}(x)

\method{TipLabels}{character}(x)
}
\arguments{
\item{x}{An object of a supported class (see Usage section).}
}
\value{
A character vector listing the tip labels for the specified object.
}
\description{
Tip labels
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
