% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_properties.R
\name{TreeIsRooted}
\alias{TreeIsRooted}
\title{Is tree rooted?}
\usage{
TreeIsRooted(tree)
}
\arguments{
\item{tree}{A phylogenetic tree of class \code{phylo}.}
}
\value{
\code{TreeIsRooted()} returns a logical specifying whether a root node is
resolved.
}
\description{
Faster alternative to of \code{ape::is.rooted}.
}
\examples{
TreeIsRooted(BalancedTree(6))

}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
