% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_properties.R
\name{TreeIsRooted}
\alias{TreeIsRooted}
\title{Is tree rooted?}
\usage{
TreeIsRooted(tree)
}
\arguments{
\item{tree}{A phylogenetic tree of class phylo.}
}
\value{
Logical specifying whether a root node is resolved.
}
\description{
Faster alternative to of \code{ape::is.rooted}.
}
\author{
Martin R. Smith
}
