% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_generation.R
\name{TreeNumber}
\alias{TreeNumber}
\alias{as.phylo.numeric}
\alias{as.TreeNumber}
\alias{as.TreeNumber.phylo}
\alias{as.TreeNumber.character}
\title{Unique integer indices for bifurcating tree topologies}
\usage{
\method{as.phylo}{numeric}(x, nTip = attr(x, "nTip"),
  tipLabels = attr(x, "tip.label"))

as.TreeNumber(x, ...)

\method{as.TreeNumber}{phylo}(x, ...)

\method{as.TreeNumber}{character}(x, nTip, tipLabels = TipLabels(nTip),
  ...)
}
\arguments{
\item{x}{Integer identifying the tree (see details).}

\item{nTip}{Integer specifying number of tips in the tree.}

\item{tipLabels}{Character vector listing the labels assigned to each tip
in a tree, perhaps obtained using \link{TipLabels}.}
}
\value{
\code{as.phylo.numeric} returns a tree of class \code{phylo}.

\code{as.TreeNumber} returns an object of class \code{TreeNumber},
which comprises numeric vector, whose elements
represent digits of the decimal integer corresponding to the tree topology
(in big endian order), with attributes \code{nTip} and \code{tip.labels}.
}
\description{
There are \code{NUnrooted(n)} unrooted trees with \emph{n} tips.
As such, each \emph{n}-tip tree can be uniquely identifyed by a non-negative
integer \emph{x} < \code{NUnrooted(n)}.
}
\details{
This integer can be converted by a tree by treating it as a mixed-base number,
with bases 1, 3, 5, 7, ... (2_n_ - 5).

Each digit of this mixed base number corresponds to a tip, and determines
the location on a growing tree to which that tip should be added.

We start with a two-tip tree, and treat 0 as the origin of the tree.\preformatted{
0 ---- 1

}

We add tip 2 by breaking an edge and inserting a node (numbered 2 + nTip - 1).
In this example, we'll work up to a six-tip tree; this node will be numbered
2 + 6 - 1 = 7.
There is only one edge on which tip 2 can be added.  Let's add node 7 and tip 2:\preformatted{
0 ---- 7 ---- 1
       |
       |
       2

}

There are now three edges on which tip 3 can be added.  Our options are:
Option 0: the edge leading to 1;
Option 1: the edge leading to 2;
Option 2: the edge leading to 7.

If we select option 1, we produce:\preformatted{
0 ---- 7 ---- 1
       |
       |
       8 ---- 2
       |
       |
       3

}

\code{1} is now the final digit of our mixed-base number

There are five places to add tip 4:
Option 0: the edge leading to 1;
Option 1: the edge leading to 2;
Option 2: the edge leading to 3;
Option 3: the edge leading to 7;
Option 4: the edge leading to 8.

If we chose option 3, then \code{3} would be the penultimate digit of our
mixed-base number

If we chose option 0 for the next two additions, we could specify this tree
with the mixed-base number 0021.  We can convert this into decimal:

0 × (1 × 3 × 5 × 9) +
0 × (1 × 3 × 5) +
3 × (1 × 3) +
1 × (1)
= 10

Note that the hyperexponential nature of tree space means that there are

2^30 unique 12-tip trees.  As integers > 2^31 are not supported by R,
numbers representing larger trees are broken into nine-digit chunks.
}
\examples{
tree <- as.phylo(10, nTip = 6)
plot(tree)
as.TreeNumber(tree)

# Larger trees:
as.numeric.phylo(BalancedTree(19))

# If > 9 digits, represent the tree number as a string.
representation <- as.TreeNumber("1234567890123", 14)
tree <- as.phylo(representation)

}
\references{
Based on a concept by John Tromp (1995)
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
