% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_shape.R
\name{TreeShape}
\alias{TreeShape}
\alias{RootedTreeShape}
\alias{RootedTreeWithShape}
\alias{NUnrootedShapes}
\alias{NRootedShapes}
\title{Integer representing shape of a tree}
\usage{
RootedTreeShape(tree)

RootedTreeWithShape(shape, nTip)

NUnrootedShapes(nTip)

NRootedShapes(nTip)
}
\arguments{
\item{tree}{A tree of class \code{\link{phylo}}.}

\item{shape}{Integer specifying shape of tree, perhaps generated by
\code{TreeShape}.}

\item{nTip}{Integer specifying number of tips.}
}
\value{
\code{TreeShape} returns an integer specifying the shape of a tree,
ignoring tip labels.

\code{TreeWithShape} returns a tree of class \code{phylo} corresponding to the
shape provided.  Tips are unlabelled.

\code{TreeShapes} returns an integer specifying the number of unique
unrooted tree shapes with \code{nTip} tips.

\code{RootedTreeShapes} returns an integer specifying the number of unique
rooted tree shapes with \code{nTip} tips.
}
\description{
Returns an integer that uniquely represents the \emph{shape} of an \emph{n}-tip
binary tree, ignoring tip labels.
}
\details{
Rooted trees are numbered working up from the root.

The root node divides \emph{n} tips into two subtrees.  The smaller subtree
may contain $a = 1, 2, ..., n/2$ tips, leaving $b = n - a$ tips in
These options are worked through in turn.

For the first shape of the smaller subtree, work through each possible shape
for the larger subtree.  Then, move to the next shape of the smaller subtree,
and work through each possible shape of the larger subtree.

Stop when the desired topology is encountered.
}
\author{
Martin R. Smith (\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
