% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeTools-package.R
\docType{package}
\name{TreeTools-package}
\alias{TreeTools}
\alias{TreeTools-package}
\title{TreeTools: Create, Modify and Analyse Phylogenetic Trees}
\description{
Core functions for the creation, modification and analysis of 
  phylogenetic trees.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ms609.github.io/TreeTools}
  \item \url{https://github.com/ms609/TreeTools}
  \item Report bugs at \url{https://github.com/ms609/TreeTools/issues}
}

}
\author{
\strong{Maintainer}: Martin R. Smith \email{martin.smith@durham.ac.uk} (0000-0001-5660-1727) [copyright holder]

Other contributors:
\itemize{
  \item Emmanuel Paradis (0000-0003-3092-2199) [copyright holder]
  \item Klaus Schliep (0000-0003-2941-0161) [copyright holder]
}

}
\keyword{internal}
