% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeTools-package.R
\docType{package}
\name{TreeTools-package}
\alias{TreeTools}
\alias{TreeTools-package}
\title{TreeTools}
\description{
'TreeTools' is an R package that provides functions for creating, modifying and
analysing phylogenetic trees.
It complements packages such as
\href{https://cran.r-project.org/package=ape}{'ape'},
\href{https://cran.r-project.org/package=phangorn}{'phangorn'} and
\href{https://cran.r-project.org/package=phytools}{'phytools'},
aiming for efficient and robust implementations of functions, typically
applied to unweighted trees (i.e. those without edge lengths).
}
\details{
Full documentation is available \href{https://ms609.github.io/TreeTools/}{online}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ms609.github.io/TreeTools/}
  \item \url{https://github.com/ms609/TreeTools/}
  \item Report bugs at \url{https://github.com/ms609/TreeTools/issues/}
}

}
\author{
\strong{Maintainer}: Martin R. Smith \email{martin.smith@durham.ac.uk} (\href{https://orcid.org/0000-0001-5660-1727}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Emmanuel Paradis (\href{https://orcid.org/0000-0003-3092-2199}{ORCID}) [copyright holder]
}

}
\keyword{internal}
