% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Splits.R
\name{xor}
\alias{xor}
\alias{xor,Splits,Splits-method}
\title{Exclusive OR operation}
\usage{
xor(x, y)

\S4method{xor}{Splits,Splits}(x, y)
}
\arguments{
\item{x, y}{Objects to be compared.}
}
\description{
Exclusive OR operation
}
\seealso{
Other Splits operations: 
\code{\link{LabelSplits}()},
\code{\link{NSplits}()},
\code{\link{NTip}()},
\code{\link{PolarizeSplits}()},
\code{\link{SplitFrequency}()},
\code{\link{Splits}},
\code{\link{SplitsInBinaryTree}()},
\code{\link{TipLabels}()},
\code{\link{TipsInSplits}()},
\code{\link{match.Splits}}
}
\concept{Splits operations}
